+++ tests/test_terminal.py
@@ -5,27 +5,33 @@ import pytest
 sys.path.insert(0, "core")
 
 from core.models import Cursor, FileSymbol, Terminal
-from core.repositories import fill_terminal
+from core.repositories import create_terminal
 
 
 class TestTerminalActions:
     @pytest.fixture
-    def empty_terminal_5_x_5(self):
+    def empty_terminal_5_x_5(self) -> Terminal:
         return Terminal(
             cursor=Cursor(x=0, y=0),
             content=[[None for _ in range(5)] for _ in range(5)],
         )
 
     @pytest.fixture
-    def empty_terminal_2_x_2(self):
+    def empty_terminal_2_x_2(self) -> Terminal:
         return Terminal(
             cursor=Cursor(x=0, y=0),
             content=[[None for _ in range(2)] for _ in range(2)],
         )
 
-    def test_perfect_terminal_fill(self, empty_terminal_5_x_5):
-        upd_terminal = fill_terminal(
-            empty_terminal_5_x_5,
+    @pytest.fixture
+    def empty_terminal_0_x_0(self) -> Terminal:
+        return Terminal(
+            cursor=Cursor(x=0, y=0),
+            content=None,
+        )
+
+    def test_perfect_terminal_fill(self, empty_terminal_5_x_5: Terminal):
+        empty_terminal_5_x_5.fill_terminal(
             [
                 FileSymbol('H', column=0, line=0),
                 FileSymbol('i', column=1, line=0),
@@ -39,7 +45,7 @@ class TestTerminalActions:
             ]
         )
 
-        assert upd_terminal.content == [
+        assert empty_terminal_5_x_5.content == [
             [
                 FileSymbol('H', column=0, line=0),
                 FileSymbol('i', column=1, line=0),
@@ -65,9 +71,8 @@ class TestTerminalActions:
             [None, None, None, None, None],
         ]
 
-    def test_multi_lines_terminal_fill(self, empty_terminal_5_x_5):
-        upd_terminal = fill_terminal(
-            empty_terminal_5_x_5,
+    def test_multi_lines_terminal_fill(self, empty_terminal_5_x_5: Terminal):
+        empty_terminal_5_x_5.fill_terminal(
             [
                 FileSymbol("I", line=0, column=0),
                 FileSymbol("'", line=1, column=0),
@@ -83,7 +88,7 @@ class TestTerminalActions:
             ]
         )
 
-        assert upd_terminal.content == [
+        assert empty_terminal_5_x_5.content == [
             [FileSymbol("I", line=0, column=0), None, None, None, None],
             [FileSymbol("'", line=1, column=0), None, None, None, None],
             [FileSymbol("m", line=2, column=0), None, None, None, None],
@@ -92,8 +97,7 @@ class TestTerminalActions:
         ]
 
     def test_overflow_terminal_fill(self, empty_terminal_2_x_2):
-        upd_terminal = fill_terminal(
-            empty_terminal_2_x_2,
+        empty_terminal_2_x_2.fill_terminal(
             [
                 FileSymbol("I", line=0, column=0),
                 FileSymbol("'", line=0, column=1),
@@ -109,7 +113,7 @@ class TestTerminalActions:
             ]
         )
 
-        assert upd_terminal.content == [
+        assert empty_terminal_2_x_2.content == [
             [
                 FileSymbol("I", line=0, column=0),
                 FileSymbol("'", line=0, column=1)
@@ -120,10 +124,32 @@ class TestTerminalActions:
             ],
         ]
 
-    def test_no_symbols_terminal_fill(self, empty_terminal_2_x_2):
-        upd_terminal = fill_terminal(empty_terminal_2_x_2, [])
+    def test_no_symbols_terminal_fill(self, empty_terminal_2_x_2: Terminal):
+        empty_terminal_2_x_2.fill_terminal([])
 
-        assert upd_terminal.content == [
+        assert empty_terminal_2_x_2.content == [
             [None, None],
             [None, None],
         ]
+
+    def test_zero_size_terminal_fill(self, empty_terminal_0_x_0: Terminal):
+        empty_terminal_0_x_0.fill_terminal(
+            [FileSymbol('H', line=0, column=0)]
+        )
+        assert empty_terminal_0_x_0.content is None
+
+    def test_create_1000_x_1000_terminal(self):
+        new_terminal = create_terminal(1000, 1000)
+        assert new_terminal.content == [[None for _ in range(1000)] for _ in range(1000)]
+
+    def test_create_0_x_0_terminal(self):
+        new_terminal = create_terminal(0, 0)
+        assert new_terminal.content is None
+
+    def test_create_1_x_0_terminal(self):
+        new_terminal = create_terminal(1, 0)
+        assert new_terminal.content is None
+
+    def test_create_0_x_1_terminal(self):
+        new_terminal = create_terminal(0, 1)
+        assert new_terminal.content is None